refactor: rename main service.ts to types.ts
This file does not hold any implementations
anymore. We can rename this to types.ts for now.
