+++ lsp/src/service/filetypes/buildspec/service.ts
@@ -1,13 +1,13 @@
 import { CompletionItem, CompletionList, Diagnostic, Hover, HoverParams, TextDocumentPositionParams } from 'vscode-languageserver'
 import { TextDocument } from 'vscode-languageserver-textdocument'
 import { YamlLanguageService } from '../../../utils/yaml/service'
-import { LanguageService } from '../../service'
+import { LanguageService } from '../../types'
 
 const SCHEMA_URL = 'https://d3rrggjwfhwld2.cloudfront.net/CodeBuild/buildspec/buildspec-standalone.schema.json'
 
-export class BuildspecService  implements LanguageService {
+export class BuildspecService implements LanguageService {
 
-    constructor(private readonly yaml: YamlLanguageService = new YamlLanguageService(SCHEMA_URL)) {}
+    constructor(private readonly yaml: YamlLanguageService = new YamlLanguageService(SCHEMA_URL)) { }
 
     completion(document: TextDocument, textDocumentPositionParams: TextDocumentPositionParams): Promise<CompletionItem[] | CompletionList> {
         return this.yaml.completion(document, textDocumentPositionParams)
